requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.23
django-appconf==1.0.6
django-bulk-signals==0.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-import-export==3.3.4
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
dnspython==2.6.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
kombu==5.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
setuptools==78.1.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
xlrd==2.0.1
xlwt==1.3.0
whitenoise==6.9.0
python-decouple==3.8
boto3==1.40.25
django-storages==1.14.6
Pillow==10.4.0
drf-api-logger==1.1.21